pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fourex"
version = "0.1.0"
description = "4X turn-based strategy sandbox for AI agents"
authors = [{ name = "Caleb Tutty", email = "caleb@mokotahi.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "python-multipart>=0.0.9",
    "websockets>=15.0.1",
    "instructor>=1.9.2",
    "openai>=1.95.1",
    "requests>=2.32.4",
    # Multi-LLM provider support
    "replicate>=0.22.0",
    "transformers>=4.30.0",
    "huggingface-hub>=0.16.0",
    # MCP server support
    "mcp>=1.0.0",
    # Observability and tracing
    "logfire>=0.20.0",
    "structlog>=23.1.0",
    # Retry and resilience
    "tenacity>=8.2.0",
    "backoff>=2.2.0",
    # Enhanced serialization
    "orjson>=3.9.0",
    # Database persistence
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --strict-config --tb=short"
markers = ["slow: marks tests as slow"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.hatch.build.targets.wheel]
packages = ["backend/src"]
